feat(map): changing default and adding optional basemaps
Map now loads with a dark basemap and has a dropdown for the user to change the basemaps from the provided options. To add additional tile layer basemaps > add an option to the html select following the format being used > add the url and attribution  to the switch statement in script.js.

Includes minor layout changes